requirements/requirements.txt
Django==1.7.5
djangorestframework==3.0.5
akismet==0.2.0